plumbing: transport/http, Add a redirect test harness
The redirect tests each grew their own pkt-line advertisement, their
own 401 challenge and their own recording handler, which is a lot of
surface for a fixture whose only job is to be uninteresting. A
handshake that fails because the advertisement is malformed fails for
a reason unrelated to the redirect it was written to pin, and it fails
that way in every test that copied it.

Put one advertisement builder, one challenge responder, one request
recorder and one handshake entry point in one place, and give the
fixture a test of its own: what `advert` builds must be an
advertisement `handshakeSmart` accepts, or everything written against
it is testing the fixture.
